+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ## [Unreleased] - ReleaseDate
 
+## [1.2.0] - 2025-08-09
+
 ### Added
 
 - Allow the registration of several network interfaces sensors.
@@ -48,7 +50,8 @@ Initial version
 
 
 <!-- next-url -->
-[Unreleased]: https://github.com/guillaumecl/mqtt-system-monitor/compare/v1.1.1...HEAD
+[Unreleased]: https://github.com/guillaumecl/mqtt-system-monitor/compare/v1.2.0...HEAD
+[1.2.0]: https://github.com/guillaumecl/mqtt-system-monitor/compare/v1.1.1...v1.2.0
 [1.1.1]: https://github.com/guillaumecl/mqtt-system-monitor/compare/v1.1.0...v1.1.1
 [1.1.0]: https://github.com/guillaumecl/mqtt-system-monitor/compare/v1.0.5...v1.1.0
 [1.0.5]: https://github.com/guillaumecl/mqtt-system-monitor/compare/v1.0.3...v1.0.5